refactor: decompose manager.ts into composable pipeline stages
Extract the monolithic NousManager turn execution (1447 lines) into
a 6-stage composable pipeline: resolve → guard → context → history →
execute → finalize. Manager retains only lifecycle, turn coordination,
and session locking (~184 lines).

New files:
- pipeline/types.ts — TurnState, RuntimeServices, shared interfaces
- pipeline/runner.ts — runStreamingPipeline + runBufferedPipeline
- pipeline/stages/{resolve,guard,context,history,execute,finalize}.ts
- pipeline/utils/build-messages.ts — message array construction
